+++ lib/features/workout/presentation/live_workout_screen.dart
@@ -3,6 +3,7 @@
 
 import 'dart:async';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:liquid_glass_widgets/liquid_glass_widgets.dart';
 import '../../../util/design_constants.dart';
 import '../../app/presentation/widgets/glass_bottom_menu.dart';
@@ -232,7 +233,7 @@ class _LiveWorkoutScreenState extends State<LiveWorkoutScreen>
     }
   }
 
-  void _onReorder(int oldIndex, int newIndex) {
+  void _onReorderItem(int oldIndex, int newIndex) {
     Provider.of<LiveWorkoutViewModel>(
       context,
       listen: false,
@@ -619,7 +620,8 @@ class _LiveWorkoutScreenState extends State<LiveWorkoutScreen>
                             child: exercises.isEmpty
                                 ? _buildEmptyState(l10n)
                                 : ReorderableListView.builder(
-                                    cacheExtent: 1500,
+                                    scrollCacheExtent:
+                                        const ScrollCacheExtent.pixels(1500.0),
                                     padding: EdgeInsets.only(
                                       bottom: (showRestBar
                                               ? 180.0
@@ -627,7 +629,7 @@ class _LiveWorkoutScreenState extends State<LiveWorkoutScreen>
                                                   .bottomContentSpacer) +
                                           MediaQuery.paddingOf(context).bottom,
                                     ),
-                                    onReorder: _onReorder,
+                                    onReorderItem: _onReorderItem,
                                     itemCount: exercises.length,
                                     itemBuilder: (context, index) {
                                       final routineExercise = exercises[index];