+++ lib/widgets/common/common.dart
@@ -11,3 +11,5 @@ export 'app_metadata_row.dart';
 export 'app_section_header.dart';
 export 'time_range_filter.dart';
 export 'long_running_operation_overlay.dart';
+export 'platform_adaptive_switch.dart';
+export 'platform_adaptive_switch_list_tile.dart';